Do it the way Mohamed said to do it
