--- graphql/resolvers/config.go
@@ -1,58 +1,0 @@
-package resolvers
-
-// This file will be automatically regenerated based on the schema, any resolver implementations
-// will be copied through when generating and any unknown code will be moved to the end.
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/evergreen-ci/evergreen"
-	gqlError "github.com/evergreen-ci/evergreen/graphql/errors"
-	restModel "github.com/evergreen-ci/evergreen/rest/model"
-)
-
-func (r *queryResolver) AwsRegions(ctx context.Context) ([]string, error) {
-	return evergreen.GetEnvironment().Settings().Providers.AWS.AllowedRegions, nil
-}
-
-func (r *queryResolver) ClientConfig(ctx context.Context) (*restModel.APIClientConfig, error) {
-	envClientConfig := evergreen.GetEnvironment().ClientConfig()
-	clientConfig := restModel.APIClientConfig{}
-	err := clientConfig.BuildFromService(*envClientConfig)
-	if err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error building APIClientConfig from service: %s", err.Error()))
-	}
-
-	return &clientConfig, nil
-}
-
-func (r *queryResolver) InstanceTypes(ctx context.Context) ([]string, error) {
-	config, err := evergreen.GetConfig()
-	if err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, "unable to retrieve server config")
-	}
-	return config.Providers.AWS.AllowedInstanceTypes, nil
-}
-
-func (r *queryResolver) SpruceConfig(ctx context.Context) (*restModel.APIAdminSettings, error) {
-	config, err := evergreen.GetConfig()
-	if err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error Fetching evergreen settings: %s", err.Error()))
-	}
-
-	spruceConfig := restModel.APIAdminSettings{}
-	err = spruceConfig.BuildFromService(config)
-	if err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error building api admin settings from service: %s", err.Error()))
-	}
-	return &spruceConfig, nil
-}
-
-func (r *queryResolver) SubnetAvailabilityZones(ctx context.Context) ([]string, error) {
-	zones := []string{}
-	for _, subnet := range evergreen.GetEnvironment().Settings().Providers.AWS.Subnets {
-		zones = append(zones, subnet.AZ)
-	}
-	return zones, nil
-}
