+++ graphql/resolvers/query.go
@@ -1,0 +1,1270 @@
+package resolvers
+
+// This file will be automatically regenerated based on the schema, any resolver implementations
+// will be copied through when generating and any unknown code will be moved to the end.
+
+import (
+	"context"
+	"fmt"
+	"sort"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/evergreen-ci/evergreen"
+	"github.com/evergreen-ci/evergreen/apimodels"
+	"github.com/evergreen-ci/evergreen/db"
+	gqlError "github.com/evergreen-ci/evergreen/graphql/errors"
+	"github.com/evergreen-ci/evergreen/graphql/generated"
+	gqlModel "github.com/evergreen-ci/evergreen/graphql/model"
+	"github.com/evergreen-ci/evergreen/graphql/resolvers/util"
+	"github.com/evergreen-ci/evergreen/model"
+	"github.com/evergreen-ci/evergreen/model/distro"
+	"github.com/evergreen-ci/evergreen/model/event"
+	"github.com/evergreen-ci/evergreen/model/host"
+	"github.com/evergreen-ci/evergreen/model/patch"
+	"github.com/evergreen-ci/evergreen/model/task"
+	"github.com/evergreen-ci/evergreen/model/testresult"
+	"github.com/evergreen-ci/evergreen/model/user"
+	"github.com/evergreen-ci/evergreen/rest/data"
+	restModel "github.com/evergreen-ci/evergreen/rest/model"
+	"github.com/evergreen-ci/evergreen/thirdparty"
+	"github.com/evergreen-ci/utility"
+	"github.com/mongodb/grip"
+	"github.com/mongodb/grip/message"
+	werrors "github.com/pkg/errors"
+)
+
+func (r *queryResolver) BbGetCreatedTickets(ctx context.Context, taskID string) ([]*thirdparty.JiraTicket, error) {
+	createdTickets, err := util.BBGetCreatedTicketsPointers(taskID)
+	if err != nil {
+		return nil, err
+	}
+
+	return createdTickets, nil
+}
+
+func (r *queryResolver) BuildBaron(ctx context.Context, taskID string, execution int) (*gqlModel.BuildBaron, error) {
+	execString := strconv.Itoa(execution)
+
+	searchReturnInfo, bbConfig, err := model.GetSearchReturnInfo(taskID, execString)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, err.Error())
+	}
+
+	return &gqlModel.BuildBaron{
+		SearchReturnInfo:        searchReturnInfo,
+		BuildBaronConfigured:    bbConfig.ProjectFound && bbConfig.SearchConfigured,
+		BbTicketCreationDefined: bbConfig.TicketCreationDefined,
+	}, nil
+}
+
+func (r *queryResolver) AwsRegions(ctx context.Context) ([]string, error) {
+	return evergreen.GetEnvironment().Settings().Providers.AWS.AllowedRegions, nil
+}
+
+func (r *queryResolver) ClientConfig(ctx context.Context) (*restModel.APIClientConfig, error) {
+	envClientConfig := evergreen.GetEnvironment().ClientConfig()
+	clientConfig := restModel.APIClientConfig{}
+	err := clientConfig.BuildFromService(*envClientConfig)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error building APIClientConfig from service: %s", err.Error()))
+	}
+
+	return &clientConfig, nil
+}
+
+func (r *queryResolver) InstanceTypes(ctx context.Context) ([]string, error) {
+	config, err := evergreen.GetConfig()
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, "unable to retrieve server config")
+	}
+	return config.Providers.AWS.AllowedInstanceTypes, nil
+}
+
+func (r *queryResolver) SpruceConfig(ctx context.Context) (*restModel.APIAdminSettings, error) {
+	config, err := evergreen.GetConfig()
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error Fetching evergreen settings: %s", err.Error()))
+	}
+
+	spruceConfig := restModel.APIAdminSettings{}
+	err = spruceConfig.BuildFromService(config)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error building api admin settings from service: %s", err.Error()))
+	}
+	return &spruceConfig, nil
+}
+
+func (r *queryResolver) SubnetAvailabilityZones(ctx context.Context) ([]string, error) {
+	zones := []string{}
+	for _, subnet := range evergreen.GetEnvironment().Settings().Providers.AWS.Subnets {
+		zones = append(zones, subnet.AZ)
+	}
+	return zones, nil
+}
+
+func (r *queryResolver) Distros(ctx context.Context, onlySpawnable bool) ([]*restModel.APIDistro, error) {
+	apiDistros := []*restModel.APIDistro{}
+
+	var distros []distro.Distro
+	if onlySpawnable {
+		d, err := distro.Find(distro.BySpawnAllowed())
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while fetching spawnable distros: %s", err.Error()))
+		}
+		distros = d
+	} else {
+		d, err := distro.FindAll()
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while fetching distros: %s", err.Error()))
+		}
+		distros = d
+	}
+	for _, d := range distros {
+		apiDistro := restModel.APIDistro{}
+		err := apiDistro.BuildFromService(d)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Unable to build APIDistro from distro: %s", err.Error()))
+		}
+		apiDistros = append(apiDistros, &apiDistro)
+	}
+	return apiDistros, nil
+}
+
+func (r *queryResolver) DistroTaskQueue(ctx context.Context, distroID string) ([]*restModel.APITaskQueueItem, error) {
+	distroQueue, err := model.LoadTaskQueue(distroID)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error getting task queue for distro %v: %v", distroID, err.Error()))
+	}
+	if distroQueue == nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("cannot find queue with distro ID `%s`", distroID))
+	}
+
+	taskQueue := []*restModel.APITaskQueueItem{}
+
+	for _, taskQueueItem := range distroQueue.Queue {
+		apiTaskQueueItem := restModel.APITaskQueueItem{}
+		err := apiTaskQueueItem.BuildFromService(taskQueueItem)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error converting task queue item db model to api model: %v", err.Error()))
+		}
+		taskQueue = append(taskQueue, &apiTaskQueueItem)
+	}
+
+	return taskQueue, nil
+}
+
+func (r *queryResolver) Host(ctx context.Context, hostID string) (*restModel.APIHost, error) {
+	host, err := host.GetHostByIdOrTagWithTask(hostID)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error Fetching host: %s", err.Error()))
+	}
+	if host == nil {
+		return nil, werrors.Errorf("unable to find host %s", hostID)
+	}
+
+	apiHost := &restModel.APIHost{}
+	err = apiHost.BuildFromService(host)
+	if err != nil || apiHost == nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error converting from host.Host to model.APIHost: %s", err.Error()))
+	}
+
+	if host.RunningTask != "" {
+		// Add the task information to the host document.
+		if err = apiHost.BuildFromService(host.RunningTaskFull); err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error converting from host.Host to model.APIHost: %s", err.Error()))
+		}
+	}
+
+	return apiHost, nil
+}
+
+func (r *queryResolver) HostEvents(ctx context.Context, hostID string, hostTag *string, limit *int, page *int) (*gqlModel.HostEvents, error) {
+	h, err := host.FindOneByIdOrTag(hostID)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error finding host %s: %s", hostID, err.Error()))
+	}
+	if h == nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Host %s not found", hostID))
+	}
+	events, err := event.FindPaginated(h.Id, h.Tag, event.AllLogCollection, *limit, *page)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error Fetching host events: %s", err.Error()))
+	}
+	// populate eventlogs pointer arrays
+	apiEventLogPointers := []*restModel.HostAPIEventLogEntry{}
+	for _, e := range events {
+		apiEventLog := restModel.HostAPIEventLogEntry{}
+		err = apiEventLog.BuildFromService(&e)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Unable to build APIEventLogEntry from EventLog: %s", err.Error()))
+		}
+		apiEventLogPointers = append(apiEventLogPointers, &apiEventLog)
+	}
+	hostevents := gqlModel.HostEvents{
+		EventLogEntries: apiEventLogPointers,
+		Count:           len(events),
+	}
+	return &hostevents, nil
+}
+
+func (r *queryResolver) Hosts(ctx context.Context, hostID *string, distroID *string, currentTaskID *string, statuses []string, startedBy *string, sortBy *gqlModel.HostSortBy, sortDir *gqlModel.SortDirection, page *int, limit *int) (*gqlModel.HostsResponse, error) {
+	hostIDParam := ""
+	if hostID != nil {
+		hostIDParam = *hostID
+	}
+	distroParam := ""
+	if distroID != nil {
+		distroParam = *distroID
+	}
+	currentTaskParam := ""
+	if currentTaskID != nil {
+		currentTaskParam = *currentTaskID
+	}
+	startedByParam := ""
+	if startedBy != nil {
+		startedByParam = *startedBy
+	}
+	sorter := host.StatusKey
+	if sortBy != nil {
+		switch *sortBy {
+		case gqlModel.HostSortByCurrentTask:
+			sorter = host.RunningTaskKey
+		case gqlModel.HostSortByDistro:
+			sorter = host.DistroKey
+		case gqlModel.HostSortByElapsed:
+			sorter = "task_full.start_time"
+		case gqlModel.HostSortByID:
+			sorter = host.IdKey
+		case gqlModel.HostSortByIDLeTime:
+			sorter = host.TotalIdleTimeKey
+		case gqlModel.HostSortByOwner:
+			sorter = host.StartedByKey
+		case gqlModel.HostSortByStatus:
+			sorter = host.StatusKey
+		case gqlModel.HostSortByUptime:
+			sorter = host.CreateTimeKey
+		default:
+			sorter = host.StatusKey
+		}
+
+	}
+	sortDirParam := 1
+	if *sortDir == gqlModel.SortDirectionDesc {
+		sortDirParam = -1
+	}
+	pageParam := 0
+	if page != nil {
+		pageParam = *page
+	}
+	limitParam := 0
+	if limit != nil {
+		limitParam = *limit
+	}
+
+	hosts, filteredHostsCount, totalHostsCount, err := host.GetPaginatedRunningHosts(hostIDParam, distroParam, currentTaskParam, statuses, startedByParam, sorter, sortDirParam, pageParam, limitParam)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error getting hosts: %s", err.Error()))
+	}
+
+	apiHosts := []*restModel.APIHost{}
+
+	for _, host := range hosts {
+		apiHost := restModel.APIHost{}
+
+		err = apiHost.BuildFromService(host)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error building API Host from Service: %s", err.Error()))
+		}
+
+		if host.RunningTask != "" {
+			// Add the task information to the host document.
+			if err = apiHost.BuildFromService(host.RunningTaskFull); err != nil {
+				return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error converting from host.Host to model.APIHost: %s", err.Error()))
+			}
+		}
+
+		apiHosts = append(apiHosts, &apiHost)
+	}
+
+	return &gqlModel.HostsResponse{
+		Hosts:              apiHosts,
+		FilteredHostsCount: filteredHostsCount,
+		TotalHostsCount:    totalHostsCount,
+	}, nil
+}
+
+func (r *queryResolver) TaskQueueDistros(ctx context.Context) ([]*gqlModel.TaskQueueDistro, error) {
+	queues, err := model.FindAllTaskQueues()
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error getting all task queues: %v", err.Error()))
+	}
+
+	distros := []*gqlModel.TaskQueueDistro{}
+
+	for _, distro := range queues {
+		numHosts, err := host.CountRunningHosts(distro.Distro)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error getting associated hosts: %s", err.Error()))
+		}
+		tqd := gqlModel.TaskQueueDistro{
+			ID:        distro.Distro,
+			TaskCount: len(distro.Queue),
+			HostCount: numHosts,
+		}
+		distros = append(distros, &tqd)
+	}
+
+	// sort distros by task count in descending order
+	sort.SliceStable(distros, func(i, j int) bool {
+		return distros[i].TaskCount > distros[j].TaskCount
+	})
+
+	return distros, nil
+}
+
+func (r *queryResolver) Patch(ctx context.Context, id string) (*restModel.APIPatch, error) {
+	patch, err := data.FindPatchById(id)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, err.Error())
+	}
+
+	if evergreen.IsFinishedPatchStatus(*patch.Status) {
+		failedAndAbortedStatuses := append(evergreen.TaskFailureStatuses, evergreen.TaskAborted)
+		opts := task.GetTasksByVersionOptions{
+			Statuses:                       failedAndAbortedStatuses,
+			FieldsToProject:                []string{task.DisplayStatusKey},
+			IncludeBaseTasks:               false,
+			IncludeBuildVariantDisplayName: false,
+		}
+		tasks, _, err := task.GetTasksByVersion(id, opts)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Could not fetch tasks for patch: %s ", err.Error()))
+		}
+
+		if len(patch.ChildPatches) > 0 {
+			for _, cp := range patch.ChildPatches {
+				// add the child patch tasks to tasks so that we can consider their status
+				childPatchTasks, _, err := task.GetTasksByVersion(*cp.Id, opts)
+				if err != nil {
+					return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Could not fetch tasks for child patch: %s ", err.Error()))
+				}
+				tasks = append(tasks, childPatchTasks...)
+			}
+		}
+		statuses := util.GetAllTaskStatuses(tasks)
+
+		// If theres an aborted task we should set the patch status to aborted if there are no other failures
+		if utility.StringSliceContains(statuses, evergreen.TaskAborted) {
+			if len(utility.StringSliceIntersection(statuses, evergreen.TaskFailureStatuses)) == 0 {
+				patch.Status = utility.ToStringPtr(evergreen.PatchAborted)
+			}
+		}
+	}
+	return patch, nil
+}
+
+func (r *queryResolver) PatchTasks(ctx context.Context, patchID string, sorts []*gqlModel.SortOrder, page *int, limit *int, statuses []string, baseStatuses []string, variant *string, taskName *string, includeEmptyActivation *bool) (*gqlModel.PatchTasks, error) {
+	pageParam := 0
+	if page != nil {
+		pageParam = *page
+	}
+	limitParam := 0
+	if limit != nil {
+		limitParam = *limit
+	}
+	variantParam := ""
+	if variant != nil {
+		variantParam = *variant
+	}
+	taskNameParam := ""
+	if taskName != nil {
+		taskNameParam = *taskName
+	}
+	var taskSorts []task.TasksSortOrder
+	if len(sorts) > 0 {
+		taskSorts = []task.TasksSortOrder{}
+		for _, singleSort := range sorts {
+			key := ""
+			switch singleSort.Key {
+			// the keys here should be the keys for the column headers of the tasks table
+			case gqlModel.TaskSortCategoryName:
+				key = task.DisplayNameKey
+			case gqlModel.TaskSortCategoryStatus:
+				key = task.DisplayStatusKey
+			case gqlModel.TaskSortCategoryBaseStatus:
+				key = task.BaseTaskStatusKey
+			case gqlModel.TaskSortCategoryVariant:
+				key = task.BuildVariantKey
+			case gqlModel.TaskSortCategoryDuration:
+				key = task.TimeTakenKey
+			default:
+				return nil, gqlError.InputValidationError.Send(ctx, fmt.Sprintf("invalid sort key: %s", singleSort.Key))
+			}
+			order := 1
+			if singleSort.Direction == gqlModel.SortDirectionDesc {
+				order = -1
+			}
+			taskSorts = append(taskSorts, task.TasksSortOrder{Key: key, Order: order})
+		}
+	}
+	v, err := model.VersionFindOne(model.VersionById(patchID).WithFields(model.VersionRequesterKey))
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while finding version with id: `%s`: %s", patchID, err.Error()))
+	}
+	if v == nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Unable to find version with id: `%s`", patchID))
+	}
+
+	opts := task.GetTasksByVersionOptions{
+		Statuses:                       util.GetValidTaskStatusesFilter(statuses),
+		BaseStatuses:                   util.GetValidTaskStatusesFilter(baseStatuses),
+		Variants:                       []string{variantParam},
+		TaskNames:                      []string{taskNameParam},
+		Page:                           pageParam,
+		Limit:                          limitParam,
+		Sorts:                          taskSorts,
+		IncludeBaseTasks:               true,
+		IncludeEmptyActivation:         utility.FromBoolPtr(includeEmptyActivation),
+		IncludeBuildVariantDisplayName: true,
+		IsMainlineCommit:               !evergreen.IsPatchRequester(v.Requester),
+	}
+	tasks, count, err := task.GetTasksByVersion(patchID, opts)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error getting patch tasks for %s: %s", patchID, err.Error()))
+	}
+
+	var apiTasks []*restModel.APITask
+	for _, t := range tasks {
+		apiTask := restModel.APITask{}
+		err := apiTask.BuildFromArgs(&t, nil)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error converting task item db model to api model: %v", err.Error()))
+		}
+		apiTasks = append(apiTasks, &apiTask)
+	}
+	patchTasks := gqlModel.PatchTasks{
+		Count: count,
+		Tasks: apiTasks,
+	}
+	return &patchTasks, nil
+}
+
+func (r *queryResolver) GithubProjectConflicts(ctx context.Context, projectID string) (*model.GithubProjectConflicts, error) {
+	pRef, err := model.FindMergedProjectRef(projectID, "", false)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error getting project: %v", err.Error()))
+	}
+	if pRef == nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("project '%s' not found", projectID))
+	}
+
+	conflicts, err := pRef.GetGithubProjectConflicts()
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error getting project conflicts: %v", err.Error()))
+	}
+	return &conflicts, nil
+}
+
+func (r *queryResolver) Project(ctx context.Context, projectID string) (*restModel.APIProjectRef, error) {
+	project, err := data.FindProjectById(projectID, true, false)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error finding project by id %s: %s", projectID, err.Error()))
+	}
+	apiProjectRef := restModel.APIProjectRef{}
+	err = apiProjectRef.BuildFromService(project)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error building APIProject from service: %s", err.Error()))
+	}
+	return &apiProjectRef, nil
+}
+
+func (r *queryResolver) Projects(ctx context.Context) ([]*gqlModel.GroupedProjects, error) {
+	allProjects, err := model.FindAllMergedTrackedProjectRefs()
+	if err != nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, err.Error())
+	}
+	// We have to iterate over the merged project refs to verify if they are enabled
+	enabledProjects := []model.ProjectRef{}
+	for _, p := range allProjects {
+		if p.IsEnabled() {
+			enabledProjects = append(enabledProjects, p)
+		}
+	}
+	groupedProjects, err := util.GroupProjects(enabledProjects, false)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error grouping project: %s", err.Error()))
+	}
+	return groupedProjects, nil
+}
+
+func (r *queryResolver) ProjectEvents(ctx context.Context, identifier string, limit *int, before *time.Time) (*gqlModel.ProjectEvents, error) {
+	timestamp := time.Now()
+	if before != nil {
+		timestamp = *before
+	}
+	events, err := data.GetProjectEventLog(identifier, timestamp, utility.FromIntPtr(limit))
+	res := &gqlModel.ProjectEvents{
+		EventLogEntries: util.GetPointerEventList(events),
+		Count:           len(events),
+	}
+	return res, err
+}
+
+func (r *queryResolver) ProjectSettings(ctx context.Context, identifier string) (*restModel.APIProjectSettings, error) {
+	projectRef, err := model.FindBranchProjectRef(identifier)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error looking in project collection: %s", err.Error()))
+	}
+	if projectRef == nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, "project doesn't exist")
+	}
+
+	res := &restModel.APIProjectSettings{
+		ProjectRef: restModel.APIProjectRef{},
+	}
+	if err = res.ProjectRef.BuildFromService(projectRef); err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error building APIProjectRef from service: %s", err.Error()))
+	}
+	if !projectRef.UseRepoSettings() {
+		// Default values so the UI understands what to do with nil values.
+		res.ProjectRef.DefaultUnsetBooleans()
+	}
+	return res, nil
+}
+
+func (r *queryResolver) RepoEvents(ctx context.Context, id string, limit *int, before *time.Time) (*gqlModel.ProjectEvents, error) {
+	timestamp := time.Now()
+	if before != nil {
+		timestamp = *before
+	}
+	events, err := data.GetEventsById(id, timestamp, utility.FromIntPtr(limit))
+	res := &gqlModel.ProjectEvents{
+		EventLogEntries: util.GetPointerEventList(events),
+		Count:           len(events),
+	}
+	return res, err
+}
+
+func (r *queryResolver) RepoSettings(ctx context.Context, id string) (*restModel.APIProjectSettings, error) {
+	repoRef, err := model.FindOneRepoRef(id)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error looking in repo collection: %s", err.Error()))
+	}
+	if repoRef == nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, "repo doesn't exist")
+	}
+
+	res := &restModel.APIProjectSettings{
+		ProjectRef: restModel.APIProjectRef{},
+	}
+	if err = res.ProjectRef.BuildFromService(repoRef.ProjectRef); err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error building APIProjectRef from service: %s", err.Error()))
+	}
+
+	// Default values so the UI understands what to do with nil values.
+	res.ProjectRef.DefaultUnsetBooleans()
+	return res, nil
+}
+
+func (r *queryResolver) ViewableProjectRefs(ctx context.Context) ([]*gqlModel.GroupedProjects, error) {
+	usr := util.MustHaveUser(ctx)
+	projectIds, err := usr.GetViewableProjectSettings()
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error getting viewable projects for '%s': '%s'", usr.DispName, err.Error()))
+	}
+
+	projects, err := model.FindProjectRefsByIds(projectIds...)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error getting projects: %v", err.Error()))
+	}
+
+	groupedProjects, err := util.GroupProjects(projects, true)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error grouping project: %s", err.Error()))
+	}
+	return groupedProjects, nil
+}
+
+func (r *queryResolver) MyHosts(ctx context.Context) ([]*restModel.APIHost, error) {
+	usr := util.MustHaveUser(ctx)
+	hosts, err := host.Find(host.ByUserWithRunningStatus(usr.Username()))
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx,
+			fmt.Sprintf("Error finding running hosts for user %s : %s", usr.Username(), err))
+	}
+	duration := time.Duration(5) * time.Minute
+	timestamp := time.Now().Add(-duration) // within last 5 minutes
+	recentlyTerminatedHosts, err := host.Find(host.ByUserRecentlyTerminated(usr.Username(), timestamp))
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx,
+			fmt.Sprintf("Error finding recently terminated hosts for user %s : %s", usr.Username(), err))
+	}
+	hosts = append(hosts, recentlyTerminatedHosts...)
+
+	var apiHosts []*restModel.APIHost
+	for _, host := range hosts {
+		apiHost := restModel.APIHost{}
+		err = apiHost.BuildFromService(host)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error building APIHost from service: %s", err.Error()))
+		}
+		apiHosts = append(apiHosts, &apiHost)
+	}
+	return apiHosts, nil
+}
+
+func (r *queryResolver) MyVolumes(ctx context.Context) ([]*restModel.APIVolume, error) {
+	usr := util.MustHaveUser(ctx)
+	volumes, err := host.FindSortedVolumesByUser(usr.Username())
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, err.Error())
+	}
+	return util.GetAPIVolumeList(volumes)
+}
+
+func (r *queryResolver) Task(ctx context.Context, taskID string, execution *int) (*restModel.APITask, error) {
+	dbTask, err := task.FindOneIdAndExecutionWithDisplayStatus(taskID, execution)
+	if err != nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, err.Error())
+	}
+	if dbTask == nil {
+		return nil, werrors.Errorf("unable to find task %s", taskID)
+	}
+	apiTask, err := util.GetAPITaskFromTask(ctx, r.sc.GetURL(), *dbTask)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, "error converting task")
+	}
+	return apiTask, err
+}
+
+func (r *queryResolver) TaskAllExecutions(ctx context.Context, taskID string) ([]*restModel.APITask, error) {
+	latestTask, err := task.FindOneId(taskID)
+	if err != nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, err.Error())
+	}
+	if latestTask == nil {
+		return nil, werrors.Errorf("unable to find task %s", taskID)
+	}
+	allTasks := []*restModel.APITask{}
+	for i := 0; i < latestTask.Execution; i++ {
+		var dbTask *task.Task
+		dbTask, err = task.FindByIdExecution(taskID, &i)
+		if err != nil {
+			return nil, gqlError.ResourceNotFound.Send(ctx, err.Error())
+		}
+		if dbTask == nil {
+			return nil, werrors.Errorf("unable to find task %s", taskID)
+		}
+		var apiTask *restModel.APITask
+		apiTask, err = util.GetAPITaskFromTask(ctx, r.sc.GetURL(), *dbTask)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, "error converting task")
+		}
+		allTasks = append(allTasks, apiTask)
+	}
+	apiTask, err := util.GetAPITaskFromTask(ctx, r.sc.GetURL(), *latestTask)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, "error converting task")
+	}
+	allTasks = append(allTasks, apiTask)
+	return allTasks, nil
+}
+
+func (r *queryResolver) TaskFiles(ctx context.Context, taskID string, execution *int) (*gqlModel.TaskFiles, error) {
+	emptyTaskFiles := gqlModel.TaskFiles{
+		FileCount:    0,
+		GroupedFiles: []*gqlModel.GroupedFiles{},
+	}
+	t, err := task.FindByIdExecution(taskID, execution)
+	if t == nil {
+		return &emptyTaskFiles, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("cannot find task with id %s", taskID))
+	}
+	if err != nil {
+		return &emptyTaskFiles, gqlError.ResourceNotFound.Send(ctx, err.Error())
+	}
+	groupedFilesList := []*gqlModel.GroupedFiles{}
+	fileCount := 0
+	if t.DisplayOnly {
+		execTasks, err := task.Find(task.ByIds(t.ExecutionTasks))
+		if err != nil {
+			return &emptyTaskFiles, gqlError.ResourceNotFound.Send(ctx, err.Error())
+		}
+		for _, execTask := range execTasks {
+			groupedFiles, err := util.GetGroupedFiles(ctx, execTask.DisplayName, execTask.Id, t.Execution)
+			if err != nil {
+				return &emptyTaskFiles, err
+			}
+			fileCount += len(groupedFiles.Files)
+			groupedFilesList = append(groupedFilesList, groupedFiles)
+		}
+	} else {
+		groupedFiles, err := util.GetGroupedFiles(ctx, t.DisplayName, taskID, t.Execution)
+		if err != nil {
+			return &emptyTaskFiles, err
+		}
+		fileCount += len(groupedFiles.Files)
+		groupedFilesList = append(groupedFilesList, groupedFiles)
+	}
+	taskFiles := gqlModel.TaskFiles{
+		FileCount:    fileCount,
+		GroupedFiles: groupedFilesList,
+	}
+	return &taskFiles, nil
+}
+
+func (r *queryResolver) TaskLogs(ctx context.Context, taskID string, execution *int) (*gqlModel.TaskLogs, error) {
+	t, err := task.FindByIdExecution(taskID, execution)
+	if err != nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("error finding task by id %s: %s", taskID, err.Error()))
+	}
+	if t == nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("cannot find task with id %s", taskID))
+	}
+	// need project to get default logger
+	p, err := data.FindProjectById(t.Project, true, true)
+	if err != nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("error finding project '%s': %s", t.Project, err.Error()))
+	}
+	if p == nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("could not find project '%s'", t.Project))
+	}
+	defaultLogger := p.DefaultLogger
+	if defaultLogger == "" {
+		defaultLogger = evergreen.GetEnvironment().Settings().LoggerConfig.DefaultLogger
+	}
+
+	// Let the individual TaskLogs resolvers handle fetching logs for the task
+	// We can avoid the overhead of fetching task logs that we will not view
+	// and we can avoid handling errors that we will not see
+	return &gqlModel.TaskLogs{TaskID: taskID, Execution: t.Execution, DefaultLogger: defaultLogger}, nil
+}
+
+func (r *queryResolver) TaskTests(ctx context.Context, taskID string, execution *int, sortCategory *gqlModel.TestSortCategory, sortDirection *gqlModel.SortDirection, page *int, limit *int, testName *string, statuses []string, groupID *string) (*gqlModel.TaskTestResult, error) {
+	dbTask, err := task.FindByIdExecution(taskID, execution)
+	if dbTask == nil || err != nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("finding task with id %s", taskID))
+	}
+	baseTask, err := dbTask.FindTaskOnBaseCommit()
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("finding base task for task %s: %s", taskID, err))
+	}
+
+	limitNum := utility.FromIntPtr(limit)
+	var sortBy, cedarSortBy string
+	if sortCategory != nil {
+		switch *sortCategory {
+		case gqlModel.TestSortCategoryStatus:
+			cedarSortBy = apimodels.CedarTestResultsSortByStatus
+			sortBy = testresult.StatusKey
+		case gqlModel.TestSortCategoryDuration:
+			cedarSortBy = apimodels.CedarTestResultsSortByDuration
+			sortBy = "duration"
+		case gqlModel.TestSortCategoryTestName:
+			cedarSortBy = apimodels.CedarTestResultsSortByTestName
+			sortBy = testresult.TestFileKey
+		case gqlModel.TestSortCategoryStartTime:
+			cedarSortBy = apimodels.CedarTestResultsSortByStart
+			sortBy = testresult.StartTimeKey
+		case gqlModel.TestSortCategoryBaseStatus:
+			cedarSortBy = apimodels.CedarTestResultsSortByBaseStatus
+			sortBy = "base_status"
+		}
+	} else if limitNum > 0 { // Don't sort TaskID if unlimited EVG-13965.
+		sortBy = testresult.TaskIDKey
+	}
+
+	if dbTask.HasCedarResults {
+		opts := apimodels.GetCedarTestResultsOptions{
+			BaseURL:      evergreen.GetEnvironment().Settings().Cedar.BaseURL,
+			TaskID:       taskID,
+			Execution:    utility.ToIntPtr(dbTask.Execution),
+			DisplayTask:  dbTask.DisplayOnly,
+			TestName:     utility.FromStringPtr(testName),
+			Statuses:     statuses,
+			GroupID:      utility.FromStringPtr(groupID),
+			SortBy:       cedarSortBy,
+			SortOrderDSC: sortDirection != nil && *sortDirection == gqlModel.SortDirectionDesc,
+			Limit:        limitNum,
+			Page:         utility.FromIntPtr(page),
+		}
+		if baseTask != nil && baseTask.HasCedarResults {
+			opts.BaseTaskID = baseTask.Id
+		}
+		cedarTestResults, err := apimodels.GetCedarTestResultsWithStatusError(ctx, opts)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("finding test results for task %s: %s", taskID, err))
+		}
+
+		apiTestResults := make([]*restModel.APITest, len(cedarTestResults.Results))
+		for i, t := range cedarTestResults.Results {
+			apiTest := &restModel.APITest{}
+			if err = apiTest.BuildFromService(t.TaskID); err != nil {
+				return nil, gqlError.InternalServerError.Send(ctx, err.Error())
+			}
+			if err = apiTest.BuildFromService(&t); err != nil {
+				return nil, gqlError.InternalServerError.Send(ctx, err.Error())
+			}
+
+			apiTestResults[i] = apiTest
+		}
+
+		return &gqlModel.TaskTestResult{
+			TestResults:       apiTestResults,
+			TotalTestCount:    cedarTestResults.Stats.TotalCount,
+			FilteredTestCount: utility.FromIntPtr(cedarTestResults.Stats.FilteredCount),
+		}, nil
+	}
+
+	baseTestStatusMap := map[string]string{}
+	if baseTask != nil {
+		baseTestResults, _ := r.sc.FindTestsByTaskId(data.FindTestsByTaskIdOpts{
+			TaskID:         baseTask.Id,
+			Execution:      baseTask.Execution,
+			ExecutionTasks: baseTask.ExecutionTasks,
+		})
+		for _, t := range baseTestResults {
+			baseTestStatusMap[t.TestFile] = t.Status
+		}
+	}
+	sortDir := 1
+	if sortDirection != nil && *sortDirection == gqlModel.SortDirectionDesc {
+		sortDir = -1
+	}
+	filteredTestResults, err := r.sc.FindTestsByTaskId(data.FindTestsByTaskIdOpts{
+		TaskID:         taskID,
+		Execution:      dbTask.Execution,
+		ExecutionTasks: dbTask.ExecutionTasks,
+		TestName:       utility.FromStringPtr(testName),
+		Statuses:       statuses,
+		SortBy:         sortBy,
+		SortDir:        sortDir,
+		GroupID:        utility.FromStringPtr(groupID),
+		Limit:          limitNum,
+		Page:           utility.FromIntPtr(page),
+	})
+	if err != nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, err.Error())
+	}
+
+	apiTestResults := make([]*restModel.APITest, len(filteredTestResults))
+	for i, t := range filteredTestResults {
+		apiTest := &restModel.APITest{}
+		if err = apiTest.BuildFromService(t.TaskID); err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, err.Error())
+		}
+		if err = apiTest.BuildFromService(&t); err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, err.Error())
+		}
+		apiTest.BaseStatus = utility.ToStringPtr(baseTestStatusMap[utility.FromStringPtr(apiTest.TestFile)])
+
+		apiTestResults[i] = apiTest
+	}
+	totalTestCount, err := task.GetTestCountByTaskIdAndFilters(taskID, "", []string{}, dbTask.Execution)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("getting total test count: %s", err))
+	}
+	filteredTestCount, err := task.GetTestCountByTaskIdAndFilters(taskID, utility.FromStringPtr(testName), statuses, dbTask.Execution)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("getting filtered test count: %s", err))
+	}
+
+	return &gqlModel.TaskTestResult{
+		TestResults:       apiTestResults,
+		TotalTestCount:    totalTestCount,
+		FilteredTestCount: filteredTestCount,
+	}, nil
+}
+
+func (r *queryResolver) TaskTestSample(ctx context.Context, tasks []string, filters []*gqlModel.TestFilter) ([]*gqlModel.TaskTestResultSample, error) {
+	const testSampleLimit = 10
+	if len(tasks) == 0 {
+		return nil, nil
+	}
+	dbTasks, err := task.FindAll(db.Query(task.ByIds(tasks)))
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error finding tasks %s: %s", tasks, err))
+	}
+	if len(dbTasks) == 0 {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("tasks %s not found", tasks))
+	}
+	testResultsToReturn := []*gqlModel.TaskTestResultSample{}
+
+	// We can assume that if one of the tasks has cedar results, all of them do.
+	if dbTasks[0].HasCedarResults {
+		failingTests := []string{}
+		for _, f := range filters {
+			failingTests = append(failingTests, f.TestName)
+		}
+
+		results, err := util.GetCedarFailedTestResultsSample(ctx, dbTasks, failingTests)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("getting test results sample: %s", err))
+		}
+		for _, r := range results {
+			tr := &gqlModel.TaskTestResultSample{
+				TaskID:                  utility.FromStringPtr(r.TaskID),
+				Execution:               r.Execution,
+				MatchingFailedTestNames: r.MatchingFailedTestNames,
+				TotalTestCount:          r.TotalFailedNames,
+			}
+			testResultsToReturn = append(testResultsToReturn, tr)
+		}
+	} else {
+		testFilters := []string{}
+		for _, f := range filters {
+			testFilters = append(testFilters, f.TestName)
+		}
+		regexFilter := strings.Join(testFilters, "|")
+		for _, t := range dbTasks {
+			filteredTestResults, err := r.sc.FindTestsByTaskId(data.FindTestsByTaskIdOpts{
+				TaskID:         t.Id,
+				Execution:      t.Execution,
+				ExecutionTasks: t.ExecutionTasks,
+				TestName:       regexFilter,
+				Statuses:       []string{evergreen.TestFailedStatus},
+				SortBy:         testresult.TaskIDKey,
+				Limit:          testSampleLimit,
+				SortDir:        1,
+				Page:           0,
+			})
+			if err != nil {
+				return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("getting test results sample: %s", err))
+			}
+			failedTestCount, err := task.GetTestCountByTaskIdAndFilters(t.Id, "", []string{evergreen.TestFailedStatus}, t.Execution)
+			if err != nil {
+				return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("getting failed test count: %s", err))
+			}
+			tr := &gqlModel.TaskTestResultSample{
+				TaskID:         t.Id,
+				Execution:      t.Execution,
+				TotalTestCount: failedTestCount,
+			}
+			matchingFailingTestNames := []string{}
+			for _, r := range filteredTestResults {
+				matchingFailingTestNames = append(matchingFailingTestNames, r.TestFile)
+			}
+			tr.MatchingFailedTestNames = matchingFailingTestNames
+			testResultsToReturn = append(testResultsToReturn, tr)
+		}
+	}
+	return testResultsToReturn, nil
+}
+
+func (r *queryResolver) MyPublicKeys(ctx context.Context) ([]*restModel.APIPubKey, error) {
+	publicKeys := util.GetMyPublicKeys(ctx)
+	return publicKeys, nil
+}
+
+func (r *queryResolver) User(ctx context.Context, userID *string) (*restModel.APIDBUser, error) {
+	usr := util.MustHaveUser(ctx)
+	var err error
+	if userID != nil {
+		usr, err = user.FindOneById(*userID)
+		if err != nil {
+			return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Error getting user from user ID: %s", err.Error()))
+		}
+		if usr == nil {
+			return nil, gqlError.ResourceNotFound.Send(ctx, "Could not find user from user ID")
+		}
+	}
+	displayName := usr.DisplayName()
+	username := usr.Username()
+	email := usr.Email()
+	user := restModel.APIDBUser{
+		DisplayName:  &displayName,
+		UserID:       &username,
+		EmailAddress: &email,
+	}
+	return &user, nil
+}
+
+func (r *queryResolver) UserConfig(ctx context.Context) (*gqlModel.UserConfig, error) {
+	usr := util.MustHaveUser(ctx)
+	settings := evergreen.GetEnvironment().Settings()
+	config := &gqlModel.UserConfig{
+		User:          usr.Username(),
+		APIKey:        usr.GetAPIKey(),
+		UIServerHost:  settings.Ui.Url,
+		APIServerHost: settings.ApiUrl + "/api",
+	}
+	return config, nil
+}
+
+func (r *queryResolver) UserSettings(ctx context.Context) (*restModel.APIUserSettings, error) {
+	usr := util.MustHaveUser(ctx)
+	userSettings := restModel.APIUserSettings{}
+	err := userSettings.BuildFromService(usr.Settings)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, err.Error())
+	}
+	return &userSettings, nil
+}
+
+func (r *queryResolver) BuildVariantsForTaskName(ctx context.Context, projectID string, taskName string) ([]*task.BuildVariantTuple, error) {
+	pid, err := model.GetIdForProject(projectID)
+	if err != nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Could not find project with id: %s", projectID))
+	}
+	repo, err := model.FindRepository(pid)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while getting repository for '%s': %s", projectID, err.Error()))
+	}
+	if repo == nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("could not find repository '%s'", projectID))
+	}
+	taskBuildVariants, err := task.FindUniqueBuildVariantNamesByTask(pid, taskName, repo.RevisionOrderNumber)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while getting build variant tasks for task '%s': %s", taskName, err.Error()))
+	}
+	if taskBuildVariants == nil {
+		return nil, nil
+	}
+	return taskBuildVariants, nil
+}
+
+func (r *queryResolver) CommitQueue(ctx context.Context, id string) (*restModel.APICommitQueue, error) {
+	commitQueue, err := data.FindCommitQueueForProject(id)
+	if err != nil {
+		if werrors.Cause(err) == err {
+			return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("error finding commit queue for %s: %s", id, err.Error()))
+		}
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error finding commit queue for %s: %s", id, err.Error()))
+	}
+	project, err := data.FindProjectById(id, true, true)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error finding project %s: %s", id, err.Error()))
+	}
+	if project.CommitQueue.Message != "" {
+		commitQueue.Message = &project.CommitQueue.Message
+	}
+	commitQueue.Owner = &project.Owner
+	commitQueue.Repo = &project.Repo
+
+	for i, item := range commitQueue.Queue {
+		patchId := ""
+		if utility.FromStringPtr(item.Version) != "" {
+			patchId = utility.FromStringPtr(item.Version)
+		} else if utility.FromStringPtr(item.PatchId) != "" {
+			patchId = utility.FromStringPtr(item.PatchId)
+		}
+
+		if patchId != "" {
+			p, err := data.FindPatchById(patchId)
+			if err != nil {
+				return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("error finding patch: %s", err.Error()))
+			}
+			commitQueue.Queue[i].Patch = p
+		}
+	}
+	return commitQueue, nil
+}
+
+func (r *queryResolver) HasVersion(ctx context.Context, id string) (bool, error) {
+	v, err := model.VersionFindOne(model.VersionById(id))
+	if err != nil {
+		return false, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error finding version %s: %s", id, err.Error()))
+	}
+	if v != nil {
+		return true, nil
+	}
+
+	if patch.IsValidId(id) {
+		p, err := patch.FindOneId(id)
+		if err != nil {
+			return false, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error finding patch %s: %s", id, err.Error()))
+		}
+		if p != nil {
+			return false, nil
+		}
+	}
+	return false, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Unable to find patch or version %s", id))
+}
+
+func (r *queryResolver) MainlineCommits(ctx context.Context, options gqlModel.MainlineCommitsOptions, buildVariantOptions *gqlModel.BuildVariantOptions) (*gqlModel.MainlineCommits, error) {
+	projectId, err := model.GetIdForProject(options.ProjectID)
+	if err != nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Could not find project with id: %s", options.ProjectID))
+	}
+	limit := model.DefaultMainlineCommitVersionLimit
+	if utility.FromIntPtr(options.Limit) != 0 {
+		limit = utility.FromIntPtr(options.Limit)
+	}
+	requesters := options.Requesters
+	if len(requesters) == 0 {
+		requesters = evergreen.SystemVersionRequesterTypes
+	}
+	opts := model.MainlineCommitVersionOptions{
+		Limit:           limit,
+		SkipOrderNumber: utility.FromIntPtr(options.SkipOrderNumber),
+		Requesters:      requesters,
+	}
+
+	versions, err := model.GetMainlineCommitVersionsWithOptions(projectId, opts)
+	if err != nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Error getting activated versions: %s", err.Error()))
+	}
+
+	var mainlineCommits gqlModel.MainlineCommits
+	matchingVersionCount := 0
+
+	// We only want to return the PrevPageOrderNumber if a user is not on the first page
+	if options.SkipOrderNumber != nil {
+		prevPageCommit, err := model.GetPreviousPageCommitOrderNumber(projectId, utility.FromIntPtr(options.SkipOrderNumber), limit, requesters)
+
+		if err != nil {
+			// This shouldn't really happen, but if it does, we should return an error and log it
+			grip.Warning(message.WrapError(err, message.Fields{
+				"message":    "Error getting most recent version",
+				"project_id": projectId,
+			}))
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error getting most recent mainline commit: %s", err.Error()))
+		}
+
+		if prevPageCommit != nil {
+			mainlineCommits.PrevPageOrderNumber = prevPageCommit
+		}
+	}
+
+	index := 0
+	versionsCheckedCount := 0
+
+	// We will loop through each version returned from GetMainlineCommitVersionsWithOptions and see if there is a commit that matches the filter parameters (if any).
+	// If there is a match, we will add it to the array of versions to be returned to the user.
+	// If there are not enough matches to satisfy our limit, we will call GetMainlineCommitVersionsWithOptions again with the next order number to check and repeat the process.
+	for matchingVersionCount < limit {
+		// If we no longer have any more versions to check break out and return what we have.
+		if len(versions) == 0 {
+			break
+		}
+		// If we have checked more versions than the MaxMainlineCommitVersionLimit then break out and return what we have.
+		if versionsCheckedCount >= model.MaxMainlineCommitVersionLimit {
+			// Return an error if we did not find any versions that match.
+			if matchingVersionCount == 0 {
+				return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Matching version not found in %d most recent versions", model.MaxMainlineCommitVersionLimit))
+			}
+			break
+		}
+		versionsCheckedCount++
+		v := versions[index]
+		apiVersion := restModel.APIVersion{}
+		err = apiVersion.BuildFromService(&v)
+		if err != nil {
+			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error building APIVersion from service: %s", err.Error()))
+		}
+
+		// If the version was created before we started caching activation status we must manually verify it and cache that value.
+		if v.Activated == nil {
+			err = util.SetVersionActivationStatus(&v)
+			if err != nil {
+				return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error setting version activation status: %s", err.Error()))
+			}
+		}
+		mainlineCommitVersion := gqlModel.MainlineCommitVersion{}
+		shouldCollapse := false
+		if !utility.FromBoolPtr(v.Activated) {
+			shouldCollapse = true
+		} else if util.IsPopulated(buildVariantOptions) && utility.FromBoolPtr(options.ShouldCollapse) {
+			opts := task.HasMatchingTasksOptions{
+				TaskNames: buildVariantOptions.Tasks,
+				Variants:  buildVariantOptions.Variants,
+				Statuses:  util.GetValidTaskStatusesFilter(buildVariantOptions.Statuses),
+			}
+			hasTasks, err := task.HasMatchingTasks(v.Id, opts)
+			if err != nil {
+				return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error checking if version has tasks: %s", err.Error()))
+			}
+			if !hasTasks {
+				shouldCollapse = true
+			}
+		}
+		// If a version matches our filter criteria we append it directly to our returned list of mainlineCommits
+		if !shouldCollapse {
+			matchingVersionCount++
+			mainlineCommits.NextPageOrderNumber = utility.ToIntPtr(v.RevisionOrderNumber)
+			mainlineCommitVersion.Version = &apiVersion
+
+		} else {
+			// If a version does not match our filter criteria roll up all the unactivated versions that are sequentially near each other into a single MainlineCommitVersion,
+			// and then append it to our returned list.
+			// If we have any versions already we should check the most recent one first otherwise create a new one
+			if len(mainlineCommits.Versions) > 0 {
+				lastMainlineCommit := mainlineCommits.Versions[len(mainlineCommits.Versions)-1]
+
+				// If the previous mainlineCommit contains rolled up unactivated versions append the latest RolledUp unactivated version
+				if lastMainlineCommit.RolledUpVersions != nil {
+					lastMainlineCommit.RolledUpVersions = append(lastMainlineCommit.RolledUpVersions, &apiVersion)
+				} else {
+					mainlineCommitVersion.RolledUpVersions = []*restModel.APIVersion{&apiVersion}
+				}
+			} else {
+				mainlineCommitVersion.RolledUpVersions = []*restModel.APIVersion{&apiVersion}
+			}
+		}
+
+		// Only add a mainlineCommit if a new one was added and it's not a modified existing RolledUpVersion
+		if mainlineCommitVersion.Version != nil || mainlineCommitVersion.RolledUpVersions != nil {
+			mainlineCommits.Versions = append(mainlineCommits.Versions, &mainlineCommitVersion)
+		}
+		index++
+		// If we have exhausted all of our versions we should fetch some more.
+		if index == len(versions) && matchingVersionCount < limit {
+			skipOrderNumber := versions[len(versions)-1].RevisionOrderNumber
+			opts := model.MainlineCommitVersionOptions{
+				Limit:           limit,
+				SkipOrderNumber: skipOrderNumber,
+				Requesters:      requesters,
+			}
+
+			versions, err = model.GetMainlineCommitVersionsWithOptions(projectId, opts)
+			if err != nil {
+				return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Error getting activated versions: %s", err.Error()))
+			}
+			index = 0
+		}
+	}
+	return &mainlineCommits, nil
+}
+
+func (r *queryResolver) TaskNamesForBuildVariant(ctx context.Context, projectID string, buildVariant string) ([]string, error) {
+	pid, err := model.GetIdForProject(projectID)
+	if err != nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Could not find project with id: %s", projectID))
+	}
+	repo, err := model.FindRepository(pid)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while getting repository for '%s': %s", projectID, err.Error()))
+	}
+	if repo == nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("could not find repository '%s'", pid))
+	}
+	buildVariantTasks, err := task.FindTaskNamesByBuildVariant(pid, buildVariant, repo.RevisionOrderNumber)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while getting tasks for '%s': %s", buildVariant, err.Error()))
+	}
+	if buildVariantTasks == nil {
+		return []string{}, nil
+	}
+	return buildVariantTasks, nil
+}
+
+func (r *queryResolver) Version(ctx context.Context, id string) (*restModel.APIVersion, error) {
+	v, err := model.VersionFindOneId(id)
+	if err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while finding version with id: `%s`: %s", id, err.Error()))
+	}
+	if v == nil {
+		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Unable to find version with id: `%s`", id))
+	}
+	apiVersion := restModel.APIVersion{}
+	if err = apiVersion.BuildFromService(v); err != nil {
+		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error building APIVersion from service for `%s`: %s", id, err.Error()))
+	}
+	return &apiVersion, nil
+}
+
+// Query returns generated.QueryResolver implementation.
+func (r *Resolver) Query() generated.QueryResolver { return &queryResolver{r} }
+
+type queryResolver struct{ *Resolver }
