+++ graphql/resolvers/version.go
@@ -24,317 +24,8 @@ import (
 	restModel "github.com/evergreen-ci/evergreen/rest/model"
 	"github.com/evergreen-ci/gimlet"
 	"github.com/evergreen-ci/utility"
-	"github.com/mongodb/grip"
-	"github.com/mongodb/grip/message"
 	werrors "github.com/pkg/errors"
 )
-
-func (r *mutationResolver) RemoveItemFromCommitQueue(ctx context.Context, commitQueueID string, issue string) (*string, error) {
-	result, err := data.CommitQueueRemoveItem(commitQueueID, issue, gimlet.GetUser(ctx).DisplayName())
-	if err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error removing item %s from commit queue %s: %s",
-			issue, commitQueueID, err.Error()))
-	}
-	if result == nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("couldn't remove item %s from commit queue %s", issue, commitQueueID))
-	}
-	return &issue, nil
-}
-
-func (r *mutationResolver) RestartVersions(ctx context.Context, versionID string, abort bool, versionsToRestart []*model.VersionToRestart) ([]*restModel.APIVersion, error) {
-	if len(versionsToRestart) == 0 {
-		return nil, gqlError.InputValidationError.Send(ctx, "No versions provided. You must provide at least one version to restart")
-	}
-	modifications := model.VersionModification{
-		Action:            evergreen.RestartAction,
-		Abort:             abort,
-		VersionsToRestart: versionsToRestart,
-	}
-	err := util.ModifyVersionHandler(ctx, versionID, modifications)
-	if err != nil {
-		return nil, err
-	}
-	versions := []*restModel.APIVersion{}
-	for _, version := range versionsToRestart {
-		if version.VersionId != nil {
-			v, versionErr := model.VersionFindOneId(*version.VersionId)
-			if versionErr != nil {
-				return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error finding version by id %s: %s", *version.VersionId, versionErr.Error()))
-			}
-			if v == nil {
-				return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("cannot find version with id %s", *version.VersionId))
-			}
-			apiVersion := restModel.APIVersion{}
-			if err = apiVersion.BuildFromService(v); err != nil {
-				return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error building APIVersion from service for `%s`: %s", *version.VersionId, err.Error()))
-			}
-			versions = append(versions, &apiVersion)
-		}
-	}
-	return versions, nil
-}
-
-func (r *queryResolver) BuildVariantsForTaskName(ctx context.Context, projectID string, taskName string) ([]*task.BuildVariantTuple, error) {
-	pid, err := model.GetIdForProject(projectID)
-	if err != nil {
-		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Could not find project with id: %s", projectID))
-	}
-	repo, err := model.FindRepository(pid)
-	if err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while getting repository for '%s': %s", projectID, err.Error()))
-	}
-	if repo == nil {
-		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("could not find repository '%s'", projectID))
-	}
-	taskBuildVariants, err := task.FindUniqueBuildVariantNamesByTask(pid, taskName, repo.RevisionOrderNumber)
-	if err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while getting build variant tasks for task '%s': %s", taskName, err.Error()))
-	}
-	if taskBuildVariants == nil {
-		return nil, nil
-	}
-	return taskBuildVariants, nil
-}
-
-func (r *queryResolver) CommitQueue(ctx context.Context, id string) (*restModel.APICommitQueue, error) {
-	commitQueue, err := data.FindCommitQueueForProject(id)
-	if err != nil {
-		if werrors.Cause(err) == err {
-			return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("error finding commit queue for %s: %s", id, err.Error()))
-		}
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error finding commit queue for %s: %s", id, err.Error()))
-	}
-	project, err := data.FindProjectById(id, true, true)
-	if err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("error finding project %s: %s", id, err.Error()))
-	}
-	if project.CommitQueue.Message != "" {
-		commitQueue.Message = &project.CommitQueue.Message
-	}
-	commitQueue.Owner = &project.Owner
-	commitQueue.Repo = &project.Repo
-
-	for i, item := range commitQueue.Queue {
-		patchId := ""
-		if utility.FromStringPtr(item.Version) != "" {
-			patchId = utility.FromStringPtr(item.Version)
-		} else if utility.FromStringPtr(item.PatchId) != "" {
-			patchId = utility.FromStringPtr(item.PatchId)
-		}
-
-		if patchId != "" {
-			p, err := data.FindPatchById(patchId)
-			if err != nil {
-				return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("error finding patch: %s", err.Error()))
-			}
-			commitQueue.Queue[i].Patch = p
-		}
-	}
-	return commitQueue, nil
-}
-
-func (r *queryResolver) HasVersion(ctx context.Context, id string) (bool, error) {
-	v, err := model.VersionFindOne(model.VersionById(id))
-	if err != nil {
-		return false, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error finding version %s: %s", id, err.Error()))
-	}
-	if v != nil {
-		return true, nil
-	}
-
-	if patch.IsValidId(id) {
-		p, err := patch.FindOneId(id)
-		if err != nil {
-			return false, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error finding patch %s: %s", id, err.Error()))
-		}
-		if p != nil {
-			return false, nil
-		}
-	}
-	return false, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Unable to find patch or version %s", id))
-}
-
-func (r *queryResolver) MainlineCommits(ctx context.Context, options gqlModel.MainlineCommitsOptions, buildVariantOptions *gqlModel.BuildVariantOptions) (*gqlModel.MainlineCommits, error) {
-	projectId, err := model.GetIdForProject(options.ProjectID)
-	if err != nil {
-		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Could not find project with id: %s", options.ProjectID))
-	}
-	limit := model.DefaultMainlineCommitVersionLimit
-	if utility.FromIntPtr(options.Limit) != 0 {
-		limit = utility.FromIntPtr(options.Limit)
-	}
-	requesters := options.Requesters
-	if len(requesters) == 0 {
-		requesters = evergreen.SystemVersionRequesterTypes
-	}
-	opts := model.MainlineCommitVersionOptions{
-		Limit:           limit,
-		SkipOrderNumber: utility.FromIntPtr(options.SkipOrderNumber),
-		Requesters:      requesters,
-	}
-
-	versions, err := model.GetMainlineCommitVersionsWithOptions(projectId, opts)
-	if err != nil {
-		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Error getting activated versions: %s", err.Error()))
-	}
-
-	var mainlineCommits gqlModel.MainlineCommits
-	matchingVersionCount := 0
-
-	// We only want to return the PrevPageOrderNumber if a user is not on the first page
-	if options.SkipOrderNumber != nil {
-		prevPageCommit, err := model.GetPreviousPageCommitOrderNumber(projectId, utility.FromIntPtr(options.SkipOrderNumber), limit, requesters)
-
-		if err != nil {
-			// This shouldn't really happen, but if it does, we should return an error and log it
-			grip.Warning(message.WrapError(err, message.Fields{
-				"message":    "Error getting most recent version",
-				"project_id": projectId,
-			}))
-			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error getting most recent mainline commit: %s", err.Error()))
-		}
-
-		if prevPageCommit != nil {
-			mainlineCommits.PrevPageOrderNumber = prevPageCommit
-		}
-	}
-
-	index := 0
-	versionsCheckedCount := 0
-
-	// We will loop through each version returned from GetMainlineCommitVersionsWithOptions and see if there is a commit that matches the filter parameters (if any).
-	// If there is a match, we will add it to the array of versions to be returned to the user.
-	// If there are not enough matches to satisfy our limit, we will call GetMainlineCommitVersionsWithOptions again with the next order number to check and repeat the process.
-	for matchingVersionCount < limit {
-		// If we no longer have any more versions to check break out and return what we have.
-		if len(versions) == 0 {
-			break
-		}
-		// If we have checked more versions than the MaxMainlineCommitVersionLimit then break out and return what we have.
-		if versionsCheckedCount >= model.MaxMainlineCommitVersionLimit {
-			// Return an error if we did not find any versions that match.
-			if matchingVersionCount == 0 {
-				return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Matching version not found in %d most recent versions", model.MaxMainlineCommitVersionLimit))
-			}
-			break
-		}
-		versionsCheckedCount++
-		v := versions[index]
-		apiVersion := restModel.APIVersion{}
-		err = apiVersion.BuildFromService(&v)
-		if err != nil {
-			return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error building APIVersion from service: %s", err.Error()))
-		}
-
-		// If the version was created before we started caching activation status we must manually verify it and cache that value.
-		if v.Activated == nil {
-			err = util.SetVersionActivationStatus(&v)
-			if err != nil {
-				return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error setting version activation status: %s", err.Error()))
-			}
-		}
-		mainlineCommitVersion := gqlModel.MainlineCommitVersion{}
-		shouldCollapse := false
-		if !utility.FromBoolPtr(v.Activated) {
-			shouldCollapse = true
-		} else if util.IsPopulated(buildVariantOptions) && utility.FromBoolPtr(options.ShouldCollapse) {
-			opts := task.HasMatchingTasksOptions{
-				TaskNames: buildVariantOptions.Tasks,
-				Variants:  buildVariantOptions.Variants,
-				Statuses:  util.GetValidTaskStatusesFilter(buildVariantOptions.Statuses),
-			}
-			hasTasks, err := task.HasMatchingTasks(v.Id, opts)
-			if err != nil {
-				return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error checking if version has tasks: %s", err.Error()))
-			}
-			if !hasTasks {
-				shouldCollapse = true
-			}
-		}
-		// If a version matches our filter criteria we append it directly to our returned list of mainlineCommits
-		if !shouldCollapse {
-			matchingVersionCount++
-			mainlineCommits.NextPageOrderNumber = utility.ToIntPtr(v.RevisionOrderNumber)
-			mainlineCommitVersion.Version = &apiVersion
-
-		} else {
-			// If a version does not match our filter criteria roll up all the unactivated versions that are sequentially near each other into a single MainlineCommitVersion,
-			// and then append it to our returned list.
-			// If we have any versions already we should check the most recent one first otherwise create a new one
-			if len(mainlineCommits.Versions) > 0 {
-				lastMainlineCommit := mainlineCommits.Versions[len(mainlineCommits.Versions)-1]
-
-				// If the previous mainlineCommit contains rolled up unactivated versions append the latest RolledUp unactivated version
-				if lastMainlineCommit.RolledUpVersions != nil {
-					lastMainlineCommit.RolledUpVersions = append(lastMainlineCommit.RolledUpVersions, &apiVersion)
-				} else {
-					mainlineCommitVersion.RolledUpVersions = []*restModel.APIVersion{&apiVersion}
-				}
-			} else {
-				mainlineCommitVersion.RolledUpVersions = []*restModel.APIVersion{&apiVersion}
-			}
-		}
-
-		// Only add a mainlineCommit if a new one was added and it's not a modified existing RolledUpVersion
-		if mainlineCommitVersion.Version != nil || mainlineCommitVersion.RolledUpVersions != nil {
-			mainlineCommits.Versions = append(mainlineCommits.Versions, &mainlineCommitVersion)
-		}
-		index++
-		// If we have exhausted all of our versions we should fetch some more.
-		if index == len(versions) && matchingVersionCount < limit {
-			skipOrderNumber := versions[len(versions)-1].RevisionOrderNumber
-			opts := model.MainlineCommitVersionOptions{
-				Limit:           limit,
-				SkipOrderNumber: skipOrderNumber,
-				Requesters:      requesters,
-			}
-
-			versions, err = model.GetMainlineCommitVersionsWithOptions(projectId, opts)
-			if err != nil {
-				return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Error getting activated versions: %s", err.Error()))
-			}
-			index = 0
-		}
-	}
-	return &mainlineCommits, nil
-}
-
-func (r *queryResolver) TaskNamesForBuildVariant(ctx context.Context, projectID string, buildVariant string) ([]string, error) {
-	pid, err := model.GetIdForProject(projectID)
-	if err != nil {
-		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Could not find project with id: %s", projectID))
-	}
-	repo, err := model.FindRepository(pid)
-	if err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while getting repository for '%s': %s", projectID, err.Error()))
-	}
-	if repo == nil {
-		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("could not find repository '%s'", pid))
-	}
-	buildVariantTasks, err := task.FindTaskNamesByBuildVariant(pid, buildVariant, repo.RevisionOrderNumber)
-	if err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while getting tasks for '%s': %s", buildVariant, err.Error()))
-	}
-	if buildVariantTasks == nil {
-		return []string{}, nil
-	}
-	return buildVariantTasks, nil
-}
-
-func (r *queryResolver) Version(ctx context.Context, id string) (*restModel.APIVersion, error) {
-	v, err := model.VersionFindOneId(id)
-	if err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error while finding version with id: `%s`: %s", id, err.Error()))
-	}
-	if v == nil {
-		return nil, gqlError.ResourceNotFound.Send(ctx, fmt.Sprintf("Unable to find version with id: `%s`", id))
-	}
-	apiVersion := restModel.APIVersion{}
-	if err = apiVersion.BuildFromService(v); err != nil {
-		return nil, gqlError.InternalServerError.Send(ctx, fmt.Sprintf("Error building APIVersion from service for `%s`: %s", id, err.Error()))
-	}
-	return &apiVersion, nil
-}
 
 func (r *versionResolver) BaseTaskStatuses(ctx context.Context, obj *restModel.APIVersion) ([]string, error) {
 	var baseVersion *model.Version
