+++ graphql/schema/types/project_subscriber.graphql
@@ -1,0 +1,48 @@
+##### TYPES #####
+"""
+ProjectSubscriber defines the subscriptions for a given Project. For example, a project could have Slack notifications
+enabled that trigger whenever any version finishes.
+"""
+type ProjectSubscriber {
+  subscriber: Subscriber!
+  type: String!
+}
+
+type Subscriber {
+  emailSubscriber: String
+  githubCheckSubscriber: GithubCheckSubscriber
+  githubPRSubscriber: GithubPRSubscriber
+  jiraCommentSubscriber: String
+  jiraIssueSubscriber: JiraIssueSubscriber
+  slackSubscriber: String
+  webhookSubscriber: WebhookSubscriber
+}
+
+type GithubPRSubscriber {
+  owner: String!
+  prNumber: Int
+  ref: String!
+  repo: String!
+}
+
+type GithubCheckSubscriber {
+  owner: String!
+  ref: String!
+  repo: String!
+}
+
+type WebhookSubscriber {
+  headers: [WebhookHeader]!
+  secret: String!
+  url: String!
+}
+
+type WebhookHeader {
+  key: String!
+  value: String!
+}
+
+type JiraIssueSubscriber {
+  issueType: String!
+  project: String!
+}
