+++ graphql/schema/types/task_queue.graphql
@@ -1,0 +1,40 @@
+"""
+Distro[] is the return value for the distros query.
+It models an environment configuration for a host.
+"""
+type Distro {
+  isVirtualWorkStation: Boolean!
+  name: String
+  user: String
+  userSpawnAllowed: Boolean
+  workDir: String
+}
+
+
+"""
+TaskQueueItem[] is the return value for the distroTaskQueue query.
+It contains information about any particular item on the task queue, such as the name of the task, the build variant of the task,
+and how long it's expected to take to finish running.
+"""
+type TaskQueueItem {
+  id: ID!
+  buildVariant: String!
+  displayName: String!
+  expectedDuration: Duration!
+  priority: Int!
+  project: String!
+  requester: TaskQueueItemType!
+  revision: String!
+  version: String!
+}
+
+"""
+TaskQueueDistro[] is the return value for the taskQueueDistros query.
+It contains information about how many tasks and hosts are running on on a particular distro.
+"""
+type TaskQueueDistro {
+  id: ID!
+  hostCount: Int!
+  taskCount: Int!
+}
+
